pyproject.toml
[tool.poetry]
name = "salaries"
version = "0.1.0"
description = ""
authors = ["Andrey Alekseev <leksus@leksus.info>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.111.0"}
pydantic = "^2.7.1"
psycopg2 = "^2.9.9"
sqlalchemy = "^2.0.30"
pyjwt = "^2.8.0"
bcrypt = "^4.1.3"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
alembic = "^1.13.1"


[tool.poetry.group.dev.dependencies]
uvicorn = {extras = ["standard"], version = "^0.29.0"}
fastapi-shell = "^0.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
